Issue a warning when auto-sens fails
Previous behavior of lockin.utils.get_ampl_phas caused an
error when the measured amplitude was too small (less than
100/65536 times the minimum sensitivity) or too large
(greater than 0.99 times the maximum sensitivity). New
behavior issues a warning and returns the offending ampl,
which allows a DMMA sweep to continue (albeit with a
warning to the user).